backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
python-dotenv==1.0.0
